requirements.txt
jsonschema==4.17.3
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
flask==2.2.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
oauthlib==3.2.2
pycparser==2.21
PyJWT==2.7.0
python-dotenv==0.21.1
requests==2.30.0
requests-oauthlib==1.3.1
urllib3==2.0.2
pymongo==4.7.1
Werkzeug==2.2.2
redis==4.4.0
python-keycloak
flasgger
flask_restful
cryptography==39.0.1
langchain
weaviate-client
openai
unstructured
unstructured[pdf]
langchain-community
langchain-anthropic